requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

# Python dependencies for the newsletter agent system
crewai>=0.16.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
apify < 3.0
nest-asyncio